:recycle: Refactor server
