+++ lib/frontend/api.ts
@@ -4,44 +4,34 @@ export interface OutlineItem {
   id: string;
 }
 
-export const fetchFileTree = async () => {
+const fetchData = async <T>(url: string, responseType: 'json' | 'text'): Promise<T | null> => {
   try {
-    const response = await fetch('http://localhost:8521/filetree');
+    const response = await fetch(url);
     if (!response.ok) {
       throw new Error(`HTTP error! status: ${response.status}`);
     }
-    const data = await response.json();
-    return data;
+    if (responseType === 'json') {
+      return await response.json();
+    } else {
+      return await response.text() as T;
+    }
   } catch (error) {
-    console.error("Error fetching file tree:", error);
-    return [];
+    console.error(`Error fetching from ${url}:`, error);
+    return null;
   }
+};
+
+export const fetchFileTree = async () => {
+  const data = await fetchData<any>('/api/filetree', 'json');
+  return data || [];
 };
 
 export const fetchContent = async (path: string) => {
-  try {
-    const response = await fetch(`http://localhost:8521/content/${path}`);
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-    const data = await response.text();
-    return data;
-  } catch (error) {
-    console.error(`Error fetching content for ${path}:`, error);
-    return "";
-  }
+  const data = await fetchData<string>(`/content/${path}`, 'text');
+  return data || "";
 };
 
 export const fetchOutline = async (filePath: string): Promise<OutlineItem[]> => {
-  try {
-    const response = await fetch(`http://localhost:8521/outline?filePath=${encodeURIComponent(filePath)}`);
-    if (!response.ok) {
-      throw new Error(`HTTP error! status: ${response.status}`);
-    }
-    const data = await response.json();
-    return data;
-  } catch (error) {
-    console.error(`Error fetching outline for ${filePath}:`, error);
-    return [];
-  }
+  const data = await fetchData<OutlineItem[]>(`/api/outline?filePath=${encodeURIComponent(filePath)}`, 'json');
+  return data || [];
 };
